move is_cert_error(); require python-requests
