+++ koji.spec
@@ -28,6 +28,7 @@ BuildArch: noarch
 Requires: python-krbV >= 1.0.13
 Requires: rpm-python
 Requires: pyOpenSSL
+Requires: python-requests
 Requires: python-urlgrabber
 Requires: python-dateutil
 BuildRequires: python
